2017.1.22 年前，修改媒体文件添加按钮

+++ app/src/main/java/com/qstech/pierglass/activity/MediaListActivity.java
@@ -3,17 +3,17 @@ package com.qstech.pierglass.activity;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.GridLayoutManager;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
@@ -24,7 +24,6 @@ import com.qstech.pierglass.adapter.MenuAdapter;
 import com.qstech.pierglass.bean.Media;
 import com.qstech.pierglass.listener.OnItemClickListener;
 import com.qstech.pierglass.utils.FileUtils;
-import com.qstech.pierglass.view.ListViewDecoration;
 import com.yanzhenjie.recyclerview.swipe.Closeable;
 import com.yanzhenjie.recyclerview.swipe.OnSwipeMenuItemClickListener;
 import com.yanzhenjie.recyclerview.swipe.SwipeMenu;
@@ -32,6 +31,7 @@ import com.yanzhenjie.recyclerview.swipe.SwipeMenuCreator;
 import com.yanzhenjie.recyclerview.swipe.SwipeMenuItem;
 import com.yanzhenjie.recyclerview.swipe.SwipeMenuRecyclerView;
 import com.yanzhenjie.recyclerview.swipe.touch.OnItemMoveListener;
+import com.yanzhenjie.recyclerview.swipe.touch.OnItemMovementListener;
 
 import java.io.BufferedWriter;
 import java.io.FileNotFoundException;
@@ -53,14 +53,24 @@ public class MediaListActivity extends AppCompatActivity {
 
     private List<Media> mMediaDatas;
 
+    private int addResourceId = R.mipmap.media_add;
+
     private MenuAdapter mMenuAdapter;
 
-    private ButtonFloat newButton;
+    //private ButtonFloat newButton;
 
     private ButtonFloat playButton;
 
+    public static final String ANDROID_RESOURCE = "android.resource://";
+    public static final String FOREWARD_SLASH = "/";
+    private static Uri resourceIdToUri(Context context, int resourceId)
+    {
+        return Uri.parse(ANDROID_RESOURCE + context.getPackageName() + FOREWARD_SLASH + resourceId);
+    }
+
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState)
+    {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.media_list);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
@@ -68,7 +78,11 @@ public class MediaListActivity extends AppCompatActivity {
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
         mContext = this;
+        Media addMedia = new Media(-1, resourceIdToUri(mContext, addResourceId), MenuAdapter.VIEW_TYPE_MENU_NONE);
+        //媒体数据
         mMediaDatas = new ArrayList<>();
+        //添加一个空数据
+        mMediaDatas.add(addMedia);
 
         SwipeMenuRecyclerView swipeMenuRecyclerView = (SwipeMenuRecyclerView) findViewById(R.id.recycler_view);
         //mLinearLayoutManager = new LinearLayoutManager(this);
@@ -76,7 +90,7 @@ public class MediaListActivity extends AppCompatActivity {
         swipeMenuRecyclerView.setLayoutManager(new GridLayoutManager(this, 3));
         swipeMenuRecyclerView.setHasFixedSize(true); //如果item够简单，高度是确定的，打开fixsize将提高性能。
         swipeMenuRecyclerView.setItemAnimator(new DefaultItemAnimator()); //设置item默认动画，加也行，不加也行。
-        swipeMenuRecyclerView.addItemDecoration(new ListViewDecoration()); //添加分割线
+        //swipeMenuRecyclerView.addItemDecoration(new ListViewDecoration()); //添加分割线
 
         //为SwipeRecyclerView的Item创建菜单就两句话
         //设置菜单创建器
@@ -90,8 +104,9 @@ public class MediaListActivity extends AppCompatActivity {
 
         swipeMenuRecyclerView.setLongPressDragEnabled(true); //开启拖拽，就这么一句话。
         swipeMenuRecyclerView.setOnItemMoveListener(onItemMoveListener); //监听拖拽，更新UI。
+        swipeMenuRecyclerView.setOnItemMovementListener(onItemMovementListener);//禁止按钮拖动删除
 
-        newButton = (ButtonFloat) findViewById(R.id.button_new);
+        /*newButton = (ButtonFloat) findViewById(R.id.button_new);
 
         newButton.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -99,8 +114,49 @@ public class MediaListActivity extends AppCompatActivity {
                 Intent intent = new Intent(mContext, MediaActivity.class);
                 startActivityForResult(intent, App.MEDIA_ADD);
             }
-        });
+        });*/
     }
+
+    /**
+     * 当Item被移动之前。
+     */
+    public OnItemMovementListener onItemMovementListener = new OnItemMovementListener() {
+        /**
+         * 当Item在移动之前，获取拖拽的方向。
+         * @param recyclerView     {@link RecyclerView}.
+         * @param targetViewHolder target ViewHolder.
+         * @return
+         */
+        @Override
+        public int onDragFlags(RecyclerView recyclerView, RecyclerView.ViewHolder targetViewHolder) {
+            // 我们让第一个不能拖拽。
+            if (targetViewHolder.getAdapterPosition() == (mMediaDatas.size()-1)) {
+                return OnItemMovementListener.INVALID;// 返回无效的方向。
+            }
+            return OnItemMovementListener.LEFT | OnItemMovementListener.RIGHT | OnItemMovementListener.UP | OnItemMovementListener.DOWN; // 可以上下左右拖拽。
+
+        }
+
+        @Override
+        public int onSwipeFlags(RecyclerView recyclerView, RecyclerView.ViewHolder targetViewHolder) {
+            // 我们让第一个不能滑动删除。
+            if (targetViewHolder.getAdapterPosition() == (mMediaDatas.size()-1)) {
+                return OnItemMovementListener.INVALID;// 返回无效的方向。
+            }
+
+            RecyclerView.LayoutManager layoutManager = recyclerView.getLayoutManager();
+            if (layoutManager instanceof LinearLayoutManager) {// 如果是LinearLayoutManager
+                LinearLayoutManager linearLayoutManager = (LinearLayoutManager) layoutManager;
+                if (linearLayoutManager.getOrientation() == LinearLayoutManager.HORIZONTAL) {// 横向的List。
+                    return OnItemMovementListener.UP | OnItemMovementListener.DOWN; // 只能上下滑动删除。
+                } else {// 竖向的List。
+                    return OnItemMovementListener.LEFT | OnItemMovementListener.RIGHT; // 只能左右滑动删除。
+                }
+            }
+            return OnItemMovementListener.INVALID;// 其它均返回无效的方向。
+        }
+    };
+
 
     /**
      * 当Item移动的时候。
@@ -109,6 +165,9 @@ public class MediaListActivity extends AppCompatActivity {
         @Override
         public boolean onItemMove(int fromPosition, int toPosition) {
             //当Item被拖拽的时候。
+            if (fromPosition == (mMediaDatas.size() - 1) || toPosition ==(mMediaDatas.size() - 1)) {
+                return false;
+            }
             Collections.swap(mMediaDatas, fromPosition, toPosition);
             mMenuAdapter.notifyItemMoved(fromPosition, toPosition);
             return true; //返回true表示处理了，返回false表示你没有处理。
@@ -135,12 +194,14 @@ public class MediaListActivity extends AppCompatActivity {
             int height = ViewGroup.LayoutParams.WRAP_CONTENT;
 
             //添加右侧的，如果不添加，则右侧不会出现菜单。
-            {
+            if (viewType == MenuAdapter.VIEW_TYPE_MENU_NONE) {// 根据Adapter的ViewType来决定菜单的样式、颜色等属性、或者是否添加菜单。
+                // Do noting
+            } else if (viewType == MenuAdapter.VIEW_TYPE_MENU_SINGLE) {// 需要添加单个菜单的Item。
                 SwipeMenuItem deleteItem = new SwipeMenuItem(mContext)
-                        .setBackgroundDrawable(R.drawable.selector_red)
-                        .setImage(R.mipmap.ic_action_delete)
-                        .setText("delete")
-                        .setTextColor(Color.WHITE)
+                        //.setBackgroundDrawable(R.drawable.selector_red)
+                        .setImage(R.mipmap.delete)
+                        //.setText("delete")
+                        //.setTextColor(Color.WHITE)
                         .setWidth(width)
                         .setHeight(0)
                         .setWeight(1);
@@ -156,6 +217,10 @@ public class MediaListActivity extends AppCompatActivity {
         public void onItemClick(int position) {
             //Toast.makeText(mContext, "第" + position + "条！", Toast.LENGTH_SHORT).show();
             Intent intent = new Intent(mContext, MediaActivity.class);
+            if (position == (mMediaDatas.size() - 1)) {
+                startActivityForResult(intent, App.MEDIA_ADD);
+                return;
+            }
             intent.putExtra("imageUri", mMediaDatas.get(position).getImageUri().toString());
             intent.putExtra("imageTime", mMediaDatas.get(position).getTime());
             intent.putExtra("position", position);
@@ -196,15 +261,17 @@ public class MediaListActivity extends AppCompatActivity {
             int position = data.getIntExtra("position", 0);
             mMediaDatas.get(position).setImageUri(imageUri);
             mMediaDatas.get(position).setTime(imageTime);
+            mMediaDatas.get(position).setViewType(MenuAdapter.VIEW_TYPE_MENU_SINGLE);
             Log.d("MediaList", "resultTime:" + mMediaDatas.get(position).getTimeString());
             mMenuAdapter.notifyDataSetChanged();
         }
         if (requestCode == App.MEDIA_ADD && resultCode == RESULT_OK) {
             Uri imageUri = Uri.parse(data.getStringExtra("imageUri"));
             int imageTime = data.getIntExtra("imageTime", 0);
-            Media media = new Media(imageTime, imageUri);
-            Log.d("MediaList", "resultTime:" + media.getTimeString());
-            mMediaDatas.add(media);
+            mMediaDatas.get(mMediaDatas.size() - 1)
+                    .setAll(imageTime, imageUri, MenuAdapter.VIEW_TYPE_MENU_SINGLE);
+            Media temp = new Media(-1, resourceIdToUri(mContext, addResourceId), MenuAdapter.VIEW_TYPE_MENU_NONE);
+            mMediaDatas.add(temp);
             mMenuAdapter.notifyDataSetChanged();
         }
     }
